Fix gas limit precision

+++ Atomex.Client.Core/Wallet/Tezos/TezosOperationFiller.cs
@@ -58,11 +58,11 @@ namespace Atomex.Wallets.Tezos
 
                     // gas limit
                     var gas = (int)tezosConfig.GasReserve
-                        + (operationResult?["consumed_milligas"]?.Value<int>() ?? 0) / 1000;
+                        + (int)(operationResult?["consumed_milligas"]?.Value<long>() ?? 0) / 1000;
 
-                    gas += metaData
+                    gas += (int)(metaData
                         ?["internal_operation_results"]
-                        ?.Sum(res => (res["result"]?["consumed_milligas"]?.Value<int>() ?? 0) / 1000) ?? 0;
+                        ?.Sum(res => (res["result"]?["consumed_milligas"]?.Value<long>() ?? 0) / 1000) ?? 0);
 
                     operation.GasLimit = gas;
